client/pyproject.toml
[project]
name = "client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "launchpad-py>=0.9.1",
    "pygame-ce>=2.5.6",
]
